Remove most of mezzanine (#1624)
This is going to be a multi-step process, where we first "empty out" all the apps that are mezzanine only, or were relevant in mezzanine context only, and then once everything looks good, do an update that removes migrations for emptied out apps, so we get rid of dumb things like one the migration files in the campaign app relying on a mezzanine.page object, and despite not having its own data anymore (because they're now done through wagtail), preventing mezzanine.pages from getting removed.

Also note that all the things that were accessible on the old /admin have been removed (but not the things living on /api) because the /admin route no longer exists. There is no access to the Django admin anymore, only the wagtail CMS admin. As such, JS and template overrides we had in place for making the Django admin look/interact nicer are now wholly irr🐘

This also introduces new migrations that move the db migration record to a safe spot, and back-edits the older migration files so that pairs of migrations that did "(X) add homepage (X+1) remove home again" are now noops (they won't kick in, but they remove app dependencies in code so that django can be told to drop apps/mezzanine dependencies)