Ensure related posts are of the same locale
Before, it could happen that related pages would be returned that had a
different locale than the reference page.

If a reference page would have a fairly unique combination of tags, it could
happen that all related pages would be different locales of the
reference page.

This is now prevented by filtering related pages to be of the same
locale as the reference page.