Move ensure related posts form save to clean
Before, when creating the blog post with the factory, there would be the
problem that ensuring the related blog posts required the locale to be
set. Wagtail only sets the local during the full_clean step for a page.
The factory does not use the full_clean method, but does call the save
method. This means when we try to ensure the related posts, the local
access would fail, because this non-nullable field would be null,
raising an exception.

To fix this issue, I moved the ensure_related_posts call from the save
method to the clean method. This still works as expected when creating
pages through the Wagtail admin, because it runs the clean method before
saving the page.

This now also allows the factory to successfully create the page. On
first save, the related posts will not be set. We need to make sure that
the factory ensures that the related posts have been set.

This is not a problem, because that is how the blog factory generate
function already worked.

We could potentially optimize this with the post-generation hooks that
factory boy provides: https://factoryboy.readthedocs.io/en/stable/reference.html#post-generation-hooks