Add Datawrapper block to base fields
This adds the block to a couple of page types streamfield definitions,
including Banner, Campaign and Default (which where required to have the
block).

This might also add the block to a couple of page types that don't need
it, but I think that is going to be an issue compared to the overhead it
would create to define these fields in a different way than it has been
so far.