Enable document links in full richtext options
This feature was previously missing which led to staff using the
external link feature to link to documents. That practice now makes it
hard to updates documents, because Wagtails typical feature of
redirecting to the latest version of the document is not working
anymore.

This commit adds the document link feature to the richtext blocks to
avoid staff from having to use the external link feature for document
links in the future.