Move profile route feature to mixin
This mixing encapulates the functionality to add a profile route to a
page (by inheritance). I tried to make the mixin composable to make it
easy to customize bits when necessary.

I have added this mixin instead of leaving the functionality on the
research author index page, because we have already discussed with the
design team that we are likely going to want profile pages on the blog
as well.