Simplify document factory
The previous implementation copied a bunch of the factories directly
from wagtail_factories. It is not clear why that was done. Maybe the
public API of wagtail_factories did not include these factories back
then?

The document factory is now simplified to inherit as much from the
original factories in wagtail_factories as possible, instead of copying
the code.