Revert "Move profile route feature to mixin"
This reverts commit 432b30fa0361bd0f828e83c4f4f0a88f60fd0720.

I actually changed my mind on this. I think the mixin is exactly the
kind of premature abstraction that has led other projects I have seen to
a state where they are hard to maintain.

This is especially true in this case where the mixin is much more
complex than the original implementation with a simple route, which was
only like 5 lines of code. Now it's 30...

Also, the only other place where I know we want something similar is the
blog. Outside of that I don't think we will create profile detail pages
in this automatic fashion.

If it turns out I am wrong and we will reuse this more often, then it
might make sense to make the mixin then. But until we have at least 3
locations where we would use it: YAGNI!