Feature/8310 research detail page (#8448)
* Add research topic and region snippets

* Unique names for region and topics

This is to avoid unnecessart duplication.

* Add research region factory

* Create regions with fake data generator

* Add research topic factory

* Add research-author relation model

This is the crutial model that creates the connection from research
detail page to the profiles.

* Extract logic to get random count of random profiles

This was previously used for blogs but it will be useful for other
factories as well.

* Generate research author relations

* Further abstract funcition to random objects of any model

* Create region and detail page relation

* Make research region and topic translatable

* Remove unique contraint from topic and region names

These fields can not be unique when the model should be translatable.
The unique contraint prevents the translation from being creatable in
the first place, even if the name would be changed in the translation
later on.

* Add topic relation

* Add research detail link

* Generate research detail links in factory

* Fix linting

* Move test file to test directory

* Use traits for make the factory more flexible

* Test URL or document (but not both) in research link

* Move document factory to own module

This feels like a bit more obvious locaiton that should facilitate
reuse.

* Make related models translatable

This was easiest with recreating the migrations in this branch

* Add additional content fields to detail page

* Add all content fields to factory

* Add method to get link url

This is necessary, so I don't need to differentiate between document and
external url in the template. Of course, the differentiation might be
interesting if I wanted to add an icon or something.

* Output content in page template

This is not styled and only a rough implementation

* Recreate migration in single file

* Remove outdated migration files

* Typo

Co-authored-by: Pomax <pomax@nihongoresources.com>

* Typo

Co-authored-by: Pomax <pomax@nihongoresources.com>

* Limit profile tagline length in factory

This is to prevent the fake text to clash with the length constraint on
the model field.

* Simplify document factory

The previous implementation copied a bunch of the factories directly
from wagtail_factories. It is not clear why that was done. Maybe the
public API of wagtail_factories did not include these factories back
then?

The document factory is now simplified to inherit as much from the
original factories in wagtail_factories as possible, instead of copying
the code.

* Import base richtext options directly

Import object instead of module because of the super long module name.

* Explicitly list translatable fields

* Renumber migration file after rebase

Co-authored-by: Pomax <pomax@nihongoresources.com>