On index show authors of correct locale
Wagtail only provides an modelinstance.localized prop. This will return the model
instance for the active locale if that exists. If not, it returns the model for
the default locale.

I want to do the same for every instance in a queryset.

Now, I could just iterate, but that would result in a new query for each and
every instance in the queryset. Not very efficient.

I want to basically do the "current if exists, default other wise" logic in the DB.

This commit is what I have come up with.

It works, but requires the ordering by locale ID. I think I need to find
a way around the ID based ordering.