Avoid locale id order requirement for queryset localization
My previous implementation of localizing a whole QuerySet was based on
the the "default" locale haing the lower ID.

This assumption (which could break if the default ID is ever changed) is
not replaced by a conditional annotation. This annotation is true if the
items locale if equal to the default and false other wise.

The rest remains the same.