Author page show alias and translated pages
Before, aliased by not fully translated pages would be missing on the
author detail page, when non-default locale was active.

This is, because the aliased pages are associated with the original
profile and not with the translated profile.

This is changed now, so that the list of displayed detail pages is based
gathered from all translations of the profile and then limited to the
current active locale.

This mean that the profile page will show alias pages that are sill
associated with the original profile and pages that are directly
associated with it self.