Create a research hub test case
I pulled in the setup code I created for #8511. Once that is
merged to main I can rebase and change the research hub index
tests to be based on the new test case class too.

The test case class is just grouping all the setup boilerplate that each
test case for the different pages in the research hub is likely to
recreate.