Replace fuzzy choice with subfactory
Before, the relation factories where using fuzzy choice. These fuzzy
choices make it harder to use the factories by them selves and were not
used anywhere anyways.

I have replaced the fuzzy choice with the more common subfactory. It
would now be possible to only create a relation and have the two models
that are made related be created automatically.