Extract queryset localization helper
I already know I need the same functionality on the library page for the
filter values of the authors. And I can already see that I will need it
for the regions and the topics too.

So, I am extracting this now, even though I don't have three localtions
yet where I need it.