Refactor library tests to test context
The test are not as targeted anymore, but they are closer to how the
page is served and works with the template. This makes more sense than
to replicate all these configurations.

The previously tested method is only used inside the class and is only
encapulating behaviour to make it easier to read and understand.

It also feels more natural to start with tests this way.