Use the HTML5 way of merging form data
Before, I was using multiple forms and merged them with hidden
fields. That was quite dirty and would leads to loads of duplication.

Luckily, HTML 5 provides a better way. We can associate input and button
elements to a form via the `form` attribute, by setting it to the id of
the form we want to merge them to.