Handle localized topic
This is the same as had already been done for the profiles.

During tree sync an alias of the pages is created. These alias pages are
of the different locales. But, the releated models are still the ones of
the default locale. Only during manual translation are the related
models also translated.

This means there can be a mix of pages in a locale, some are
translated some are only aliased. We need to make sure that when we are
filtering for the localized topic that we show pages that are associated
with that localized version of the topic or the with equivalent topic in
the default locale.