Reorder the larger layout components for side by side
The tricky part here was to create a layout that is handling the non-JS
mobile layout as well as the desktop layout. The main pain point here
was that I want the filter box to be below the results (to have the
results above the filter, because it seems like a usbility nightmare to
have them below this super long list of filters).

I can reorder the elements with flex box to an extend, but, I also need
to align the upper ends of the search box surrounding and the results
box. That was possible with a negative margin at the top.