Create more research detail pages fewer related models
This is to be better able to work on and demo the listing work.

Also, this prevents the duplicate association of authors, topics and
regiond with the research detail pages. Before, every reserach detail
pages had at least two of each, which is unrealistic. By preventing the
default subfactory to run during the fake data loading, this duplication
can be avoided and more sensible numbers of authors, topics and regions
are associated with the research detail pages.