Enable corase search ranking
I just learned that the Postgres search internally only supports 4 boost
levels / weights for search results. The Wagtail search automatically
groups the used boost values and translates them into the 4 weights that
Postgres allows. This is currently not in the docs for Wagtail search
and should probably be added.

However, with this new knowledge, I was able to activate the ranking
between the different fields -- at least roughly. Because we only have 4
levels but 5 fields, I had to group some of the fields using the same
boost values. That also required the tests to be adjusted, because
within the group we have no control over the ranking.

Another caveat is that Wagtail uses the boost values project wide. That
means any boostvalue used in any model (even third-party) needs to be
considered when grouping the boost values into the 4 weight groups. It
is possible to find the existing boost value to search weight mapping,
by stating a Django shell (`manage.py shell`):

```
>>> from wagtail.search.backends.database.postgres import weights
>>> weights.BOOSTS_WEIGHTS
[(10.0, 'A'), (7, 'B'), (4, 'C'), (1, 'D')]``
```

This is how I found the values I should be using for the four groups.