Move space between cards to list/grid fragment
The gutters between the card had previously been defined on the cards
themselves. This would have made the reuse of the cards in different
layouts more complex.

With this commit, the gutters/gaps are now defined in the entry_cards
fragement. (This is also the only place that the card templates are used.)
This should now make it easier to reuse the cards in different layouts
than the "entry_cards" way.