Fix: load list items instead of list
I think this was broken when I added the search card style. The result
was that a whole new list of search results was added as a block instead
of the existing list being extended. Depending on the page size, this
could lead to odd gaps in the design.

The templates are now restructured to be more modular. This makes it
possible to render a whole list, a single list item or a set of list
items (but without the surrounding list) and re-instates the previous
behaviour (extending the existing list instead of adding a whole new
list).