Add non link fallback to blog author fragments
Turns out, the cards are used in other contexts, where the blog index
page is not necessarily available in the context. Instead of seeking all
these locations out now, I am adding a fallback to the card so that it
works also without it.

To do so, I had to reorganise the logic a bit, because I was getting
weird gaps around the comma and Django's spaceless does not work between
text and tags, but only between tags themselves.