Add CI step for PyRight
The CI step is allowed to fail. This means you won't really be notified
if there is an issue in the type checking. You would need to check the
logs of the CI step manually to see it.

If we decide in the future that we want to make typing necessary, we
could remove the allowed failure. We would probably want to spend
sometime before to fix the existing 300+ issues. Maybe simply by
ignoring them.