Move byline out of link
The rebases turned the lower section of the blog card into a link. The
author names are also links. This caused issues, because links are not
allowed to be nested.

Since the blog card template is the only template that extends the
generic_card I just moved to byline out of that link link.