Ensure load more routes actually called
The previous test was so weak that I did not notice that the routes I
was hoping to trigger did not really fire. Some other route was
responding with the plain search page. I still don't know which one it
is, but some route was catching what I thought should have gone to the
load more route.

Turns out, I needed to properly end the search route regex patterns with
`$`. Now they are actually invoked.