Override get_entries instead of get_all_entries
It seems counter intuitive to override get_all_entries and do the
filtering in there. Especially because it was creating special cases for
filtered pages in which it was really returning ALL entries.

The override is now change to be of get_entries. This is also already
handling the "filtered" case in the super class.