Remove unnecessary __all__ definition
The definition of the `__all__` list is only necessary when we want to
limit the names that are imported with a `from x import *` statement.

Since we are importing all names that are available in the module, we
don't need that list.

Also, if we would need the list, we would need to use strings instead of
objects as the items of the list.

See also:
https://docs.python.org/3/tutorial/modules.html?highlight=__all__#importing-from-a-package