Remove unnecessary translation module
The module is outdated accoriding to a comment in the module in itself.
The module did import another module `modeltranslation` which does not
exist. That means `from networkapi.wagtaipages import translation` was
crashing with `ModuleNotFoundError: No module named 'modeltranslation'`.
That error and a search through the codebase show that the `translation`
module is not used anywhere.

Therefore, I am remove that module (this also gets rid of two type
errors it was causeing).

Closes #8468