Create and use the create_admin management command
When setting up the `pyrun` task I introduced another layer through
which all the strings are passing. That broke the `createsuperuser` task
that was again trying to pass raw Python strings as a shell string...
which was already inside of the tasks.py and had to pass through the new
pyrun methods...

This was just one layer of string escaping too many.

I have now moved the actual creation of the admin to a management
command and now we only need to run that command with a the task.