Set up basic template rendering for the article
Because of the template inheritance and the PNI menu this requires
particular data to be present in the tempalate context.

I think this could probably be refactored into a template tag to make
things a bit easier and not require each page type to pass this
information into the context.