Rename find_editorial_content_index to get_editorial_content_index
Using `get` is just a way more common pattern than `find`. I used `find`
to express that the function will return `None` instead of raising an
exception...
but there are already so many methods and functions in Django and
Wagtail that do return `None` that are start with `get` that this pattern
makes no sense to introduce. This will just be confusing...