Make retrieval of buyersguide homepage more robust
Before, the retrieval was just pulling the parent of the parent of the
article page.

Now, I am grabing the first BuyersGuidePage in the list of ancestors of
the article page. This should be a bit more robust as it gurantees the
type of the page and won't just return any page that happens to sit in
the ancestors at the right spot.