Fix orderable relations
The previous implementation had a serious flaw in that the order from
the orderable was lost. Somehow I had missed that.

Now, the order from the orderable is preserved. To ensure that we can
find the order we need the reverse relation from the "related item" to
the orderable to have a name. Otherwise we can not access the relation.