Set sort order in factory
Before, the sort order of the related articles was not defined. This
led to the order changing when the page is first saved in the Wagtail
admin. Probably because this is the first time that Wagtail fills in the
missing sort order value.

Now the sort order is properly defined in the factory and is consistent
from before to after the page being manually saved in the Wagtail admin.