Implement get_related_articles directly on category
The QuerySet method was already giving me a bunch of grief in the past
and now it breaks preview. It might be simpler to just create methods on
the models that want to give easy access to the related models through
dedicated methods.

This commit implements get_related_articles in such a fashion that it
can be turned into a utility function in the next step.