9151 buyersguide product and category related articles (#9216)
* Add related articles orderable to product category

* Fix failing test by adding missing data

Because of how the snippet formset is constructed by Wagtail, we need to
provide data for the related fields. I think it's not a great way to
test since we have to replicate too much of Wagtail's internals..

* Use form data helper to ensure all required fields are present

* Add test helper method to wrap for form data

* Add related article to categories during fake data gen

* Remove unnecessary block override

* Display current categories related articles

This is not ideal yet, because the category can be changed with JS. So I
will need to add the related articles for every category to the
template.

* Add related articles for all categories to template

The article sections that are not associate with the current category
are hidden. The showing still needs to be implemented in JS when the
frontend work is done.

* Fix orderable relations

The previous implementation had a serious flaw in that the order from
the orderable was lost. Somehow I had missed that.

Now, the order from the orderable is preserved. To ensure that we can
find the order we need the reverse relation from the "related item" to
the orderable to have a name. Otherwise we can not access the relation.

* Add a test around the related articles

* Deparate attempt to fix the duplication...

I did not work. I don't know why. I am going nuts...

* formatting

* removing previous attempted fix

* removing whitespace

* added related articles to product pages + tests + factory

* now using secondary and primary articles, however, it would be best to not have each model define that function

* updated editorial content index page to have related articles

* updated article index page to include related articles, now just need to add tests

* added tests for getting primary and secondary articles

* Cat page: split primary and secondary

The primary and secondary related articles are displayed in different
sections of the page. This adopts a pattern closer to the design to
avoid confusion during QA.

* Prod page: split primary and secondary related articles

* Fix unsemantic heading level

* Fix format

* Move less important classes lower in module

* Remove not needed heading override

There is not acceptance criteria requesting this feature. The template is also not using it.

* Fix import sort order

* Fix test

* Reactivate the failing test

* Fix failing test by returning a list instead of QuerySet

* Reduce the number of queries required to access the related items

* Remove outdated imports

* fixup! Remove not needed heading override

* Fix test assertions

The assertions where based on the idea that the related items are
returned as a queryset. Since that attempt failed I has to switch to a
list, which also works. There is a different assert method though.

* Fix type annotations

* Set sort order in factory

Before, the sort order of the related articles was not defined. This
led to the order changing when the page is first saved in the Wagtail
admin. Probably because this is the first time that Wagtail fills in the
missing sort order value.

Now the sort order is properly defined in the factory and is consistent
from before to after the page being manually saved in the Wagtail admin.

* Fix factory order

* Move relation below the main class

* Remove outdated comment

* Implement get_related_articles directly on category

The QuerySet method was already giving me a bunch of grief in the past
and now it breaks preview. It might be simpler to just create methods on
the models that want to give easy access to the related models through
dedicated methods.

This commit implements get_related_articles in such a fashion that it
can be turned into a utility function in the next step.

* Create utility function to get_related_items

* Use related_items utility on product page

* Add get_author_profiles method to article page

* Use utility in article page get_related_articles

* Add get_content_categories method to article page

* Use utility for related articles on content index

* Add utility methods to buyersguide homepage

* Fix typing issues

* Fix format

* Move orderables module to the utility app

This make sense because the utitity function is not limited to the
wagtailpages app.

* Remove the useless queryset definition

This was not any good at anything and only cause problems...

* Fix indent of home template

Co-authored-by: Daniel Miranda <manieldiranda@gmail.com>