Align call to action box with sibling elements
I have remove the padding that was creating room for the image overlap
because in the designs, box is always lining up with the siblings. It's
not the top of the icon that is lining up, but the top of the box. The
overlap is desired. We need to make sure where we use the box that there
is enough room for the overlapping icon.