Remove image alt text because image is decorative
When using a screen reader on the realted product card it woul read the
alt text (which was the product name), then the company name and then
the produt name. So you would get the product name twice.

Since the images only represent the product and the title is listed
right below, I would say they qualify as decorative. Therefore, I am
removing the alt text.