Simplify the layout by using grid
I stuck to flexbox before, because I was trying to keep the centering of
the last line. But, I was not able to create a positioning of the
newsletter box that did not jump all over the place.

I think the grid version now is acceptable, although I am still not able
to make sure a larger number of product cards is placed before the
newsletter when hiding elements in the list.