Remove unused page template `bannered_campaign_page.html`
There are no references to this template anywhere in the repo. I found
this because this would have been a location to use `unlocalize` and I
wanted to open a test page while making the change. Since this template
is unused, I can not test it. And since it is unused, it also seems
unnecessary to keep around. If we miss it in the future, we can restore
it from git.