Cleanup the topic filtering
We have not logged any of the events in a while. Also, the issue that we
tried to log was likely linked to the filtering in Python. Now we are
filtering on the database level and should avoid that issue all
together.

Also, no logs relating to the issue could be found. So it seems this was
not an issue anymore (even with the in Python filtering).