Add option to get items fragment from page serve method
Before, there was only a special route to request the items fragment.
That would mean I would have to duplicate that behaviour for any other
route that will be added.

Now, the page serve method can differentiate between an HTMX request and
a normal one. If it's from HTMX, we only send the fragment back.