Make featured advice article synced field
Before it was not defined. That meant that the field would not be
updated at all after the first creation of a translation of the buyer's
guide homepage.

Now it will always be synced to use the same page as the english/default
page. In a next step we need to make sure we link to the
translated/localized variants.