Fetch before push
Git is rejecting the push because it is not aware of the contents of the
remote repository. We need to "fetch first".

See also:
https://github.com/mozilla/foundation.mozilla.org/actions/runs/3527634554/jobs/5917171535#step:6:9

Typically you would have fetched or pulled when you first got the code,
but this is not the case here were we have not interacted with the
remote before. Thus we must fetch.