Update flake8 excluded files
This should make it easier to run flake8 on all Python files but don't
check certain directories. Before we needed to specify the files and
directories to check more explicitly.

Using the config file is easier to reproduce in different useage
scenarios than having to always specify the same file patterns.