Ensure buyer's guide CTA uses 100% width
It is the default behaviour for a div to fill 100% of the parent width
and gets it's height from the content. When the parent uses flex box,
the default width behaviour is overridden and the width is also based
on the content.

The product grid uses a flex wrapper for the CTA to stretch it of the
height of the other grid elements. This disables the default 100% width.
We need to make the width behaviour explicit so it still uses 100% of
the width in the flex container.

This issue becomes apparent when the CTA content is not long enough to
fill the with of the 2 gird spaces.