Break conditional tags in attrs to multiple Lines
We sometmes use conditional tags (`{% if ... %}`) in HTML tag
attributes, specifically `class`, to apply styling based on template
variables. To make this more readable and ensure that we have whitespace
around the class names, we need to use the formattin option to break the
attirbutes into multiple lines.

The whitespace around that class names is important so that Tailwind
will properly read the class names.