Remove primary and secondary related articles from aritcle page
We are only going to have one set of related articles defined on the
page from now. This set should be displayed as the "what to read next"
articles on the article page.

In a follow-up ticket (#9874) we will create a new stream field block
that will be used instead of the primary related articles to implement
the related articles that are embedded in the body content of the page.