Fix author markup and newsletter spacing
The different elements in the bottom of the page were using different
approaches to define their width based on the wrappers and the nesting
of `tw-container`, `tw-row`, and `tw-px-4` and the use of the bootstrap
`col` classes. This is now more consistent.

Also, the different elements not define a top margin for consistent
spacing of the elements.