Add slug as id on section and href in heading anchor
The anchor does not really provide much value at this point. I have only
added it to make it similar to how the richtext heading anchors work.

But, instead of setting the id on the anchor I am setting it on the
section. This makes sure the whole section is visible when linked to it
and that the image is not cut off based on the arbitrary offset in the
`fragment-id` class.