Double check pushed branch in if statement
We only want to run Percy CI on push to main and when any PR is
approved. This should have been the case before already, but the if
statement was reading a bit confusing, it made it look like we would
run on any push. The if statement is changed now to double check that we
are on main (which is also defined in the push branches).