Remove None default value for id lists arguments
I am not sure why the previous style was used. It seemed more
complicated than necessary. Maybe for consistency with the sort option?
Not sure. It did not have any benefit. Not the empty list (from the
contructor not the literal) is used as the default value. This happened
immediately at the beginning of the method anyhow.