Remove meaningless alt text
The alt text should describe the content of the image and not be a
static text that is the same regardless of the image. It is meaningless
for screenreader users to hear there is an image if we don't tell them
what the image is. If we want to use and alt text here, we should add
one to the image that describes it's content.