11643 fix html validation errors on product page (#11648)
* Only phrasing content allowed in p

But, when rendering richtext the have other p tags and all the content
is also wrapped in a div. So that structure was not allowed.

* Use section for section like elements

* Add alt text for external link icon

* Reuse external link icon

Making a fragment for this because the accessibility text is long.

* Empty alt on decorative icon

* Sections needs to have headings

Using a div instead because the children are already sections and it's
closer to the other tabs markup.